centralise css creation
