+++ dev/start-dev-server.js
@@ -1,26 +1,17 @@
 #!/usr/bin/env node
 
-const fs = require('fs');
 const path = require('path');
-const glob = require('glob');
 const webpack = require('webpack');
-const sass = require('node-sass');
 const chokidar = require('chokidar');
-const autoprefixer = require('autoprefixer');
-const postcss = require('postcss');
-const pxtorem = require('postcss-pxtorem');
-const mkdirp = require('mkdirp');
-const pify = require('pify');
 const browserSync = require('browser-sync').create();
 
 const webpackBundler = webpack(require('../webpack.config.js')());
 const bsConfig = require('./bs-config');
-const target = path.join(__dirname, '../', 'static', 'target');
+
 const src = path.join(__dirname, '../', 'static', 'src');
 const sassDir = path.resolve(src, 'stylesheets');
-const { sassSettings, browserslist, remifications } = require('./css-settings');
-const sassRenderP = pify(sass.render);
-const writeFileP = pify(fs.writeFile);
+
+const compileSass = require('../tools/compile-css');
 
 let INITIAL_BUNDLE = true;
 
@@ -48,30 +39,4 @@ webpackBundler.watch({
 
 chokidar
     .watch(`${sassDir}/**/*.scss`)
-    .on('change', () => {
-        const getFiles = query => glob.sync(path.resolve(sassDir, query));
-
-        getFiles('!(_|ie9|old-ie)*.scss').map((filePath) => {
-            const dest = path.resolve(target, 'stylesheets', path.relative(sassDir, filePath).replace('scss', 'css'));
-            const sassOptions = Object.assign({
-                file: filePath,
-                outFile: dest,
-            }, sassSettings);
-            const postcssPlugins = [
-                autoprefixer({ browsers: browserslist }),
-                pxtorem(remifications)
-            ];
-
-            return sassRenderP(sassOptions)
-                .then(result => postcss(postcssPlugins).process(result.css.toString()))
-                .then(result => writeFileP(dest, result.css))
-                .then(() => browserSync.reload('**/*.css'))
-                .catch(err => {
-                    browserSync.sockets.emit('fullscreen:message', {
-                        title: 'Sass Error:',
-                        body: err.formatted,
-                        timeout: 100000,
-                    });
-                });
-        });
-    });
+    .on('change', () => compileSass('!(_|ie9|old-ie)*.scss'));
