+++ tools/__tasks__/compile/css/sass.js
@@ -1,42 +1,4 @@
-const fs = require('fs');
-const path = require('path');
-
-const sass = require('node-sass');
-const mkdirp = require('mkdirp');
-const glob = require('glob');
-const postcss = require('postcss');
-const autoprefixer = require('autoprefixer');
-const pxtorem = require('postcss-pxtorem');
-const pify = require('pify');
-
-const { src, target } = require('../../config').paths;
-const { sassSettings, browserslist, remifications } = require('../../../../dev/css-settings');
-
-const sassDir = path.resolve(src, 'stylesheets');
-
-const sassRenderP = pify(sass.render);
-const writeFileP = pify(fs.writeFile);
-const getFiles = query => glob.sync(path.resolve(sassDir, query));
-
-const compile = (query, { browsers, remify = true }) => Promise.all(
-    getFiles(query).map((filePath) => {
-        const dest = path.resolve(target, 'stylesheets', path.relative(sassDir, filePath).replace('scss', 'css'));
-        const sassOptions = Object.assign({
-            file: filePath,
-            outFile: dest,
-        }, sassSettings);
-
-        const postcssPlugins = [autoprefixer({ browsers })];
-        if (remify) {
-            postcssPlugins.push(pxtorem(remifications));
-        }
-
-        mkdirp.sync(path.parse(dest).dir);
-        return sassRenderP(sassOptions)
-            .then(result => postcss(postcssPlugins).process(result.css.toString()))
-            .then(result => writeFileP(dest, result.css));
-    })
-);
+const compile = require('../../../compile-css');
 
 module.exports = {
     description: 'Compile Sass',
@@ -53,14 +15,10 @@ module.exports = {
         }),
     }, {
         description: 'Modern',
-        task: () => compile('!(_|ie9|old-ie)*.scss', {
-            browsers: browserslist,
-        }),
+        task: () => compile('!(_|ie9|old-ie)*.scss'),
     }, {
         description: 'Inline',
-        task: () => compile('inline/*.scss', {
-            browsers: browserslist,
-        }),
+        task: () => compile('inline/*.scss'),
     }],
     concurrent: true,
 };
